+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.10.1] - 2023-07--3
+
+### Fixed
+- firmware: Adding a proxy `Firmware.device` property to ensure retro-compatibility with older
+            `Firmware` usages. Previous change was breaking a lot of code.
+
 ## [1.10.0] - 2023-07-03
 
 ### Changed